backend/pyproject.toml
[tool.poetry]
name = "teepon"
version = "0.1.0"
description = "Agile Loop hackaton"
authors = ["Teepon Team"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "*"
flask = "*"
numpy = "*"
pymongo = "*"
bson = "*"

pyyaml="6.0.1"
spotipy="2.23.0"
langchain="0.0.315"
colorama="0.4.6"
tiktoken="0.5.1"
openai="0.28.1"
sqlalchemy="*"
flask-sqlalchemy="*"
mysqlclient = "^2.2.4"
flask_migrate="*"

exceptiongroup = "*"
async-timeout = "*"
poetry="*"

[tool.poetry.dev-dependencies]
autopep8 = "*"
pytest = "*"
pytest-watch = "*"
model-bakery = "*"
locust = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

